# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProfileRotationMode =
  (typeof ProfileRotationMode)[keyof typeof ProfileRotationMode];

export const ProfileRotationMode = {
  "round-robin": "round-robin",
  manual: "manual",
} as const;

export interface ApiKey {
  id: number;
  profileId: number;
  /** @nullable */
  label: string | null;
  keyValue: string;
  disabled: boolean;
  createdAt: string;
}

export interface Profile {
  id: number;
  name: string;
  targetUrl: string;
  currentKeyIndex: number;
  rotationMode: ProfileRotationMode;
  createdAt: string;
  updatedAt: string;
  keys: ApiKey[];
  modelCount: number;
  enabledModelCount: number;
}

export type ProfileModelSource =
  (typeof ProfileModelSource)[keyof typeof ProfileModelSource];

export const ProfileModelSource = {
  fetched: "fetched",
  manual: "manual",
} as const;

export interface ProfileModel {
  id: number;
  profileId: number;
  modelName: string;
  source: ProfileModelSource;
  disabled: boolean;
  createdAt: string;
}

export interface CreateProfileBody {
  name: string;
  targetUrl: string;
}

export type UpdateProfileBodyRotationMode =
  (typeof UpdateProfileBodyRotationMode)[keyof typeof UpdateProfileBodyRotationMode];

export const UpdateProfileBodyRotationMode = {
  "round-robin": "round-robin",
  manual: "manual",
} as const;

export interface UpdateProfileBody {
  name?: string;
  targetUrl?: string;
  rotationMode?: UpdateProfileBodyRotationMode;
}

export interface AddKeyBody {
  keyValue: string;
  /** @nullable */
  label?: string | null;
}

export interface LayerAccessKey {
  id: number;
  /** @nullable */
  label: string | null;
  keyValue: string;
  createdAt: string;
}

export interface UpdateProfileKeyBody {
  keyValue?: string;
  /** @nullable */
  label?: string | null;
  disabled?: boolean;
}

export interface AddModelBody {
  modelName: string;
}

export interface UpdateModelBody {
  disabled?: boolean;
}

export interface RefreshModelsResponse {
  added: number;
  removed: number;
  total: number;
  models: ProfileModel[];
}

export interface CreateAccessKeyBody {
  /** @nullable */
  label?: string | null;
}

export interface ErrorResponse {
  error: string;
}