# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: profiles
    description: Proxy profile management
  - name: keys
    description: API key management
  - name: models
    description: Per-profile model management
  - name: access-keys
    description: Layer-level access key management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List all profiles
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create a profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "201":
          description: Created profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProfile
      tags: [profiles]
      summary: Delete a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /profiles/{id}/keys:
    get:
      operationId: listProfileKeys
      tags: [keys]
      summary: List API keys for a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: addProfileKey
      tags: [keys]
      summary: Add an API key to a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddKeyBody"
      responses:
        "201":
          description: Added key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /profiles/{id}/keys/{keyId}:
    patch:
      operationId: updateProfileKey
      tags: [keys]
      summary: Update an API key's value, label, or disabled state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: keyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileKeyBody"
      responses:
        "200":
          description: Updated key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProfileKey
      tags: [keys]
      summary: Remove an API key from a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: keyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /profiles/{id}/rotate-key:
    post:
      operationId: rotateProfileKey
      tags: [keys]
      summary: Manually rotate to next API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rotated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /profiles/{id}/models:
    get:
      operationId: listProfileModels
      tags: [models]
      summary: List models for a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfileModel"
    post:
      operationId: addProfileModel
      tags: [models]
      summary: Add a manual model to a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddModelBody"
      responses:
        "201":
          description: Added model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileModel"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /profiles/{id}/models/{modelId}:
    patch:
      operationId: updateProfileModel
      tags: [models]
      summary: Update a model (toggle disabled)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: modelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModelBody"
      responses:
        "200":
          description: Updated model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileModel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProfileModel
      tags: [models]
      summary: Delete a manual model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: modelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /profiles/{id}/models/refresh:
    post:
      operationId: refreshProfileModels
      tags: [models]
      summary: Refresh models from upstream provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Refreshed model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshModelsResponse"
        "400":
          description: Upstream error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /access-keys:
    get:
      operationId: listAccessKeys
      tags: [access-keys]
      summary: List all layer access keys
      responses:
        "200":
          description: List of access keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LayerAccessKey"
    post:
      operationId: createAccessKey
      tags: [access-keys]
      summary: Create a new layer access key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccessKeyBody"
      responses:
        "201":
          description: Created access key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LayerAccessKey"

  /access-keys/{id}:
    delete:
      operationId: deleteAccessKey
      tags: [access-keys]
      summary: Delete a layer access key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Profile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        targetUrl:
          type: string
        currentKeyIndex:
          type: integer
        rotationMode:
          type: string
          enum: [round-robin, manual]
        createdAt:
          type: string
        updatedAt:
          type: string
        keys:
          type: array
          items:
            $ref: "#/components/schemas/ApiKey"
        modelCount:
          type: integer
        enabledModelCount:
          type: integer
      required:
        - id
        - name
        - targetUrl
        - currentKeyIndex
        - rotationMode
        - createdAt
        - updatedAt
        - keys
        - modelCount
        - enabledModelCount

    ApiKey:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        label:
          type: ["string", "null"]
        keyValue:
          type: string
        disabled:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - profileId
        - label
        - keyValue
        - disabled
        - createdAt

    ProfileModel:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        modelName:
          type: string
        source:
          type: string
          enum: [fetched, manual]
        disabled:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - profileId
        - modelName
        - source
        - disabled
        - createdAt

    CreateProfileBody:
      type: object
      properties:
        name:
          type: string
        targetUrl:
          type: string
      required:
        - name
        - targetUrl

    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        targetUrl:
          type: string
        rotationMode:
          type: string
          enum: [round-robin, manual]

    AddKeyBody:
      type: object
      properties:
        keyValue:
          type: string
        label:
          type: ["string", "null"]
      required:
        - keyValue

    LayerAccessKey:
      type: object
      properties:
        id:
          type: integer
        label:
          type: ["string", "null"]
        keyValue:
          type: string
        createdAt:
          type: string
      required:
        - id
        - label
        - keyValue
        - createdAt

    UpdateProfileKeyBody:
      type: object
      properties:
        keyValue:
          type: string
        label:
          type: ["string", "null"]
        disabled:
          type: boolean
      required: []

    AddModelBody:
      type: object
      properties:
        modelName:
          type: string
      required:
        - modelName

    UpdateModelBody:
      type: object
      properties:
        disabled:
          type: boolean
      required: []

    RefreshModelsResponse:
      type: object
      properties:
        added:
          type: integer
        removed:
          type: integer
        total:
          type: integer
        models:
          type: array
          items:
            $ref: "#/components/schemas/ProfileModel"
      required:
        - added
        - removed
        - total
        - models

    CreateAccessKeyBody:
      type: object
      properties:
        label:
          type: ["string", "null"]
      required: []

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error