# lib/api-zod/src/generated/types/profileModel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileModelSource } from "./profileModelSource";

export interface ProfileModel {
  id: number;
  profileId: number;
  modelName: string;
  source: ProfileModelSource;
  disabled: boolean;
  createdAt: string;
}